+++ src/device/woceilinglight.ts
@@ -2,9 +2,8 @@
  *
  * woceilinglight.ts: Switchbot BLE API registration.
  */
-import type * as Noble from '@stoprocent/noble'
-
 import type { ceilingLightProServiceData, ceilingLightServiceData } from '../types/bledevicestatus.js'
+import type { NobleTypes } from '../types/types.js'
 
 import { Buffer } from 'node:buffer'
 
@@ -112,7 +111,7 @@ export class WoCeilingLight extends SwitchbotDevice {
     return data
   }
 
-  constructor(peripheral: Noble.Peripheral, noble: typeof Noble) {
+  constructor(peripheral: NobleTypes['peripheral'], noble: NobleTypes['noble']) {
     super(peripheral, noble)
   }